COM-2612 move completion certificates
